insomnia edits
created and added bookmark image; included it on the authorAccount page only so far. created an loginError page and updated the /login routes for author and reader. created a logout page and updated the /logout route. updates to the document titles on a couple of pages. added css to the input.css file to change default look of scrollbar.